Move open period logic to groupopenperiod.py
