+++ src/sentry/issues/status_change.py
@@ -13,8 +13,9 @@ from sentry.integrations.tasks.kick_off_status_syncs import kick_off_status_sync
 from sentry.issues.ignored import IGNORED_CONDITION_FIELDS
 from sentry.issues.ongoing import TRANSITION_AFTER_DAYS
 from sentry.models.activity import Activity
-from sentry.models.group import Group, GroupStatus, update_group_open_period
+from sentry.models.group import Group, GroupStatus
 from sentry.models.grouphistory import record_group_history_from_activity_type
+from sentry.models.groupopenperiod import update_group_open_period
 from sentry.models.groupsubscription import GroupSubscription
 from sentry.models.project import Project
 from sentry.notifications.types import GroupSubscriptionReason
